pyproject.toml
[project]
name = "apify-incoming-linkedin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "rich>=14.2.0",
    "textual>=1.0.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
]
